Name validation
The idea is validating strings from any characters that could not belong to a human name. The function checks for at least one or more alpha characters of the specified locale at the beginning, allowing none or only one space or dash (-) but not both, following by at least one more alpha character. The validation does not cover first and last name altogether as instances of the above algorithm are common to names across the globe (last name examples: Van Buren, De Marco, etc...). It is recommended to be used for first, middle and last name seperately.

Also another thing I tried, I tried to remove the need of providing a locale by attempting all the locales through a for loop until the string hits the correct one. I find this kind of functioning much more convinient for applications intended for international use, and also it prevents locale-crossing strings (example: ΑβγWxyZ) but Im not quite there yet. If someone wishes to develop this idea, my code is at the bottom of the file.

If the maintainers found this useful, I would love to see this in a future release (not interested in credits, just like to improving useful stuff)